Create Version-2
For this project, we will be expanding on Version 1.

The box company you work for in an attempt to bring in more customers has decided to offer fluctuating prices based on the cubic feet of the package. 
The new rates are:
- Small Packages under 15 cf = $1.50 per cf
- Medium Packages 15 cf to 45 cf = $2.50 per cf
- Large Packages over 45 cf = $3.00 per cf

Requirements:
- Create a variable for each of the three package sizes, initiating each to the price per cubic foot.
- To determine which price to use, use either if statements or switch statements to accomplish this task.

Next Task:
The box company has decided to put a limit on the size of the packages. It will no longer ship packages larger than 65 cubic feet. The company wants to cashiers alerted to packages that are too large during the dimensions collection.

Requirements:
Create a loop around to dimensions collection of your program so it checks to make sure the package is under the more than 65 cubic feet limit and to give the cashier a chance to change the dimensions in case of input error before it gets processed for a price.