Create Version-4
Now let's further expand on Version 3. You will program in the ability to have multiple containers on the same transaction.

This new code has to be its own function and there is no limit on how many containers can be shipped on a transaction.

From this point forward, all supporting functions must be prototyped.